move few methods to test class
